+++ _examples/logger.go
@@ -40,21 +40,24 @@ func main() {
     logger.Warn("this is a warn message!")
     logger.Error("this is an error message!")
 
-    // NewLogger creates a new Logger holder with given level and handlers.
+    // NewLoggerWithoutEncoder creates a new Logger holder with given level and handlers.
     // As you know, file also can be written, just replace os.Stdout with your file!
     // A logger is made of level and handlers, so we provide some handlers for use, see logit.Handler.
     // This method is the most original way to create a logger for use.
-    logger = logit.NewLogger(logit.DebugLevel, logit.NewDefaultHandler(os.Stdout, logit.NewDefaultEncoder("2006/01/02 15:04:05")))
+    logger = logit.NewLogger(logit.DebugLevel, logit.NewDefaultHandler(os.Stdout, "2006/01/02 15:04:05"))
     logger.Info("What time is it now?")
+
+    // For convenience, we provide a register mechanism and you can use handlers like this:
+    logger = logit.NewLogger(logit.DebugLevel, logit.HandlerOf("default"))
+    logger.Info("What handler is it now?")
 
     // NewLoggerFrom creates a new Logger holder with given config.
     // The config has all the things to create a logger, such as level.
     // We provide some encoders: default encoder and json encoder.
     // See logit.Encoder to check more information.
     logger = logit.NewLoggerFrom(logit.Config{
-        Level:   logit.DebugLevel,
-        Writer:  os.Stdout,
-        Encoder: logit.NewJsonEncoder("2006/01/02 15:04:05", false),
+        Level:    logit.DebugLevel,
+        Handlers: []logit.Handler{logit.NewJsonHandler(os.Stdout, "")},
     })
     logger.Info("I am a json log!")